+++ packages/datasource-sql/CHANGELOG.md
@@ -1,3 +1,14 @@
+## @forestadmin/datasource-sql [1.7.43](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.7.42...@forestadmin/datasource-sql@1.7.43) (2023-12-12)
+
+
+
+
+
+### Dependencies
+
+* **@forestadmin/datasource-sequelize:** upgraded to 1.5.26
+* **@forestadmin/datasource-toolkit:** upgraded to 1.29.1
+
 ## @forestadmin/datasource-sql [1.7.42](https://github.com/ForestAdmin/agent-nodejs/compare/@forestadmin/datasource-sql@1.7.41...@forestadmin/datasource-sql@1.7.42) (2023-11-24)